# [opt] Update repcodes less often